+++ packages/desktop_webview_window/windows/web_view.cc
@@ -137,21 +137,20 @@ void WebView::OnWebviewControllerCreated() {
   webview_->add_NavigationStarting(
       Callback<ICoreWebView2NavigationStartingEventHandler>(
           [this](ICoreWebView2 *sender, ICoreWebView2NavigationStartingEventArgs *args) {
+            method_channel_->InvokeMethod(
+                "onNavigationStarted",
+                std::make_unique<flutter::EncodableValue>(flutter::EncodableMap{
+                    {flutter::EncodableValue("id"), flutter::EncodableValue(web_view_id_)},
+                }));
             LPWSTR uri;
             args->get_Uri(&uri);
-            BOOL is_user_initiated;
-            args->get_IsUserInitiated(&is_user_initiated);
-            BOOL is_redirect;
-            args->get_IsRedirected(&is_redirect);
-            auto method_args = flutter::EncodableMap{
-                {flutter::EncodableValue("id"), flutter::EncodableValue(web_view_id_)},
-                {flutter::EncodableValue("url"), flutter::EncodableValue(wide_to_utf8(std::wstring(uri)))},
-                {flutter::EncodableValue("isUserInitiated"), flutter::EncodableValue(is_user_initiated == TRUE)},
-                {flutter::EncodableValue("isRedirect"), flutter::EncodableValue(is_redirect == TRUE)},
-            };
-            method_channel_->InvokeMethod("onNavigationStarted",
-                                          std::make_unique<flutter::EncodableValue>(method_args));
-            return 0;
+            method_channel_->InvokeMethod(
+                "onUrlRequested",
+                std::make_unique<flutter::EncodableValue>(flutter::EncodableMap{
+                    {flutter::EncodableValue("id"), flutter::EncodableValue(web_view_id_)},
+                    {flutter::EncodableValue("url"), flutter::EncodableValue(wide_to_utf8(std::wstring(uri)))},
+                }));
+            return S_OK;
           }
       ).Get(), nullptr);
   webview_->add_NavigationCompleted(
@@ -258,7 +257,6 @@ bool WebView::CanGoForward() const {
   }
   return false;
 }
-
 
 WebView::~WebView() {
   if (webview_) {